Dang ky va Dang nhap JWT

+++ app/services/contact.service.js
@@ -19,6 +19,7 @@ class ContactService {
       favorite: payload.favorite === "true" || payload.favorite === true,
       group: group,
       avatar: payload.avatar,
+      userId: payload.userId,
     };
 
     Object.keys(contact).forEach(
@@ -42,21 +43,24 @@ class ContactService {
     return await cursor.toArray();
   }
 
-  async findByName(name) {
+  async findByName(userId, name) {
     return await this.find({
+      userId: userId,
       name: { $regex: new RegExp(new RegExp(name)), $options: "i" },
     });
   }
 
-  async findById(id) {
+  async findById(userId, id) {
     return await this.Contact.findOne({
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      userId: userId,
     });
   }
 
-  async update(id, payload) {
+  async update(userId, id, payload) {
     const filter = {
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      userId: userId,
     };
     const update = this.extractContactData(payload);
     const result = await this.Contact.findOneAndUpdate(
@@ -67,19 +71,20 @@ class ContactService {
     return result;
   }
 
-  async delete(id) {
+  async delete(userId, id) {
     const result = await this.Contact.findOneAndDelete({
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      userId: userId,
     });
     return result;
   }
 
-  async findFavorite() {
-    return await this.find({ favorite: true });
+  async findFavorite(userId) {
+    return await this.find({ userId: userId, favorite: true });
   }
 
-  async deleteAll() {
-    const result = await this.Contact.deleteMany({});
+  async deleteAll(userId) {
+    const result = await this.Contact.deleteMany({ userId: userId });
     return result.deletedCount;
   }
 }
